update-README.md
# WordPress

This is a WordPress repository configured to run on the [Pagely platform](https://pagely.com/).

![alt text] (https://f.cloud.github.com/assets/69169/2290250/c35d867a-a017-11e3-86be-cd7c5bf3ff9b.gif  "FHS-Pagely)



## Getting Started

![alt text] https://f.cloud.github.com/assets/69169/2290250/c35d867a-a017-11e3-86be-cd7c5bf3ff9b.gif  "FHS-Pagely)

Pagely is website platform optimized and configured to run high performance sites with an amazing developer workflow. There is built-in support for features such as Varnish, Redis, Apache Solr, New Relic, Nginx, PHP-FPM, MySQL, PhantomJS and more. 


## Get files from Pagely

![alt text](https://raw.githubusercontent.com/Blueprint-Marketing/fhs-1/master/pagely.png "Get files from Pagely")


### 1. Spin-up a site

If you do not yet have a Pagely account, you can create one for free. Once you've verified your email address, you will be able to add sites from your dashboard. Choose "WordPress" to use this distribution.

![alt text](https://d3vv6lp55qjaqc.cloudfront.net/items/02052s1o0J0G2y2a1U1g/Image%202019-01-23%20at%205.22.31%20PM.png "Get files from Pagely")

### 2. Load up the site

When the spin-up process is complete, you will be redirected to the site's dashboard. Click on the link under the site's name to access the Dev environment.

![alt](http://i.imgur.com/2wjCj9j.png?1, '')

### 3. Run the WordPress installer

How about the WordPress database config screen? No need to worry about database connection information as that is taken care of in the background. The only step that you need to complete is the site information and the installation process will be complete.

We will post more information about how this works but we recommend developers take a look at `wp-config.php` to get an understanding.

![alt] https://github.com/Blueprint-Marketing/fhs-1/blob/master/home.jpg, 'Home')

If you would like to keep a separate set of configuration for local development, you can use a file called `wp-config-local.php`, which is already in our .gitignore file.

### 4. Enjoy!
![alt](https://github.com/Blueprint-Marketing/fhs-1/blob/master/home.jpg?1, '')

![alt](https://github.com/Blueprint-Marketing/fhs-1/blob/master/home.jpg, "FHS")